Refactor PostgreSQLDialectExceptionMapper (#20631)
